atualização

+++ app/Database/Migrations/2026-08-23-195500_AddAboutCtaBtnToProducts.php
@@ -8,21 +8,25 @@ class AddAboutCtaBtnToProducts extends Migration
 {
     public function up()
     {
-        $fields = [
-            'about_cta_btn' => [
-                'type' => 'VARCHAR',
-                'constraint' => '255',
-                'null' => true,
-                'default' => 'Garantir Meu Frasco no WhatsApp',
-                'after' => 'about_content'
-            ],
-        ];
+        if (! $this->db->fieldExists('about_cta_btn', 'products')) {
+            $fields = [
+                'about_cta_btn' => [
+                    'type' => 'VARCHAR',
+                    'constraint' => '255',
+                    'null' => true,
+                    'default' => 'Garantir Meu Frasco no WhatsApp',
+                    'after' => 'about_content'
+                ],
+            ];
 
-        $this->forge->addColumn('products', $fields);
+            $this->forge->addColumn('products', $fields);
+        }
     }
 
     public function down()
     {
-        $this->forge->dropColumn('products', 'about_cta_btn');
+        if ($this->db->fieldExists('about_cta_btn', 'products')) {
+            $this->forge->dropColumn('products', 'about_cta_btn');
+        }
     }
 }
